Rename: 컨트랙트 call 함수 위치 변경
