+++ client/src/styles/Loading.module.css
@@ -31,7 +31,7 @@
   animation-name: loading;
   animation-duration: 3s;
   animation-iteration-count: infinite;
-  opacity: 0;
+  opacity: 1;
 }
 
 .loadingText span:nth-of-type(1) {
